requirements.txt
apache-airflow>=2.7.1
python-dotenv>=1.0.0
pyodbc>=4.0.39
pandas>=2.0.0
pika>=1.3.0  # For RabbitMQ
pytz>=2023.3
